Revert remove unnecessary int to avoid diff pollute
This reverts commit f8f43a20263ce21bf511aa913a809d9228d0562b.

+++ pymatgen/ext/matproj_legacy.py
@@ -1660,5 +1660,5 @@ def get_chunks(sequence: Sequence[Any], size=1):
     Returns:
         list[Sequence[Any]]: input sequence in chunks of length size.
     """
-    chunks = math.ceil(len(sequence) / float(size))
+    chunks = int(math.ceil(len(sequence) / float(size)))
     return [sequence[i * size : (i + 1) * size] for i in range(chunks)]
